RunState: Implement maybe_forward_network_time()
Not sure about the name, but this is a new central place where
the decision when it's okay to update network time is. It checks for
pseudo_realtime and packet source existence as well as packet
source idleness.

There's a new const that allows to completely disable
network_time_forwarding.